Reorganize how details are handled in DetailsView
The intention of this change is to make the adding of any new content
into DetailsView easier, almost automatic.
From now, new state is to be added, read the value from db, and add
value to the 'details' array.
Those steps will add new row with given data to the screen, as they are
now rendered each row for each of the 'details' entry.